+++ el16ttb/lib/Matrix/Matrix.h
@@ -5,7 +5,7 @@ using namespace std;
 #include <vector>
 #include <math.h>
 
-typedef std::vector<std::vector<double> > matrix_type;
+typedef std::vector<std::vector<float> > matrix_type;
 
 struct Shape{
   int width;
@@ -20,30 +20,30 @@ public:
   Matrix(matrix_type m);
   Matrix(Shape shape);
   Matrix(int width, int height);
-  Matrix(int width, int height, double* m);
+  Matrix(int width, int height, float* m);
   //~Matrix();
 
   Shape shape;
 
   void print_matrix();
 
-  void set(int w, int h, double v);
-  double get(int w, int h);
+  void set(int w, int h, float v);
+  float get(int w, int h);
 
   Matrix transpose();
-  double dot(Matrix other);
+  float dot(Matrix other);
   Matrix cross(Matrix other);
-  double distance(Matrix other);
-  double homogDistance(Matrix other);
+  float distance(Matrix other);
+  float homogDistance(Matrix other);
   Matrix copy();
 
-  //std::vector<double> operator[](int index); // getter only, not setter
+  //std::vector<float> operator[](int index); // getter only, not setter
   Matrix operator+(Matrix p);
   Matrix operator-();
   Matrix operator-(Matrix p);
   Matrix operator*(Matrix p);
-  Matrix operator*(double x);
-  Matrix operator/(double x);
+  Matrix operator*(float x);
+  Matrix operator/(float x);
 private:
   void Setup(matrix_type m);
   matrix_type _matrix;
